requirements.txt
streamlit
fastapi
uvicorn
python-dotenv
requests
roboflow
python-multipart
pydantic